% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.h.R
\name{ahp}
\alias{ahp}
\title{Analytic Hierarchy Process}
\usage{
ahp(data, vars, itemmat = TRUE, weights = FALSE, cir = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{itemmat}{.}

\item{weights}{.}

\item{cir}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$itemmat} \tab \tab \tab \tab \tab a table \cr
  \code{results$weights} \tab \tab \tab \tab \tab a table \cr
  \code{results$cir} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$itemmat$asDF}

\code{as.data.frame(results$itemmat)}
}
\description{
Analytic Hierarchy Process
}
