% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsurvey.h.R
\name{ahpsurvey}
\alias{ahpsurvey}
\title{AHP for survey data}
\usage{
ahpsurvey(
  data,
  vars,
  atts = "cult, fam, house, jobs, trans",
  plot1 = TRUE,
  ap = TRUE,
  method = "geomatric",
  aj = TRUE,
  method1 = "geomatric"
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{atts}{.}

\item{plot1}{.}

\item{ap}{.}

\item{method}{.}

\item{aj}{.}

\item{method1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$ap} \tab \tab \tab \tab \tab a table \cr
  \code{results$aj} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$ap$asDF}

\code{as.data.frame(results$ap)}
}
\description{
AHP for survey data
}
