% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.h.R
\name{cc}
\alias{cc}
\title{Canonical Correlation}
\usage{
cc(
  data,
  set1,
  set2,
  cor = TRUE,
  test = FALSE,
  coef = FALSE,
  load = FALSE,
  cross = FALSE,
  redun = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{.}

\item{set1}{.}

\item{set2}{.}

\item{cor}{.}

\item{test}{.}

\item{coef}{.}

\item{load}{.}

\item{cross}{.}

\item{redun}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$cor} \tab \tab \tab \tab \tab a table \cr
  \code{results$test} \tab \tab \tab \tab \tab a table \cr
  \code{results$coef} \tab \tab \tab \tab \tab a table \cr
  \code{results$load} \tab \tab \tab \tab \tab a table \cr
  \code{results$cross} \tab \tab \tab \tab \tab a table \cr
  \code{results$redun} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$scores} \tab \tab \tab \tab \tab an output \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$cor$asDF}

\code{as.data.frame(results$cor)}
}
\description{
Canonical Correlation
}
