% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtheory.h.R
\name{gtheory}
\alias{gtheory}
\title{Generalizability Theory}
\usage{
gtheory(
  data,
  dep,
  id,
  sub,
  ng = 3,
  facet,
  g = FALSE,
  d = FALSE,
  formula = "value ~ (1 | subject) + (1 | task) + (1 | rater:task) + (1 | subject:task)",
  mea = FALSE,
  item = FALSE,
  formula1 = "Score ~ (1 | Person) + (1 | Item)",
  t = "uni",
  itemd = FALSE,
  comp = FALSE,
  bm = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{id}{.}

\item{sub}{.}

\item{ng}{.}

\item{facet}{.}

\item{g}{.}

\item{d}{.}

\item{formula}{.}

\item{mea}{.}

\item{item}{.}

\item{formula1}{.}

\item{t}{.}

\item{itemd}{.}

\item{comp}{.}

\item{bm}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$g} \tab \tab \tab \tab \tab a table \cr
  \code{results$d} \tab \tab \tab \tab \tab a table \cr
  \code{results$mea} \tab \tab \tab \tab \tab a table \cr
  \code{results$item} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$comp} \tab \tab \tab \tab \tab a table \cr
  \code{results$bm} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$g$asDF}

\code{as.data.frame(results$g)}
}
\description{
Generalizability Theory
}
