% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.h.R
\name{map}
\alias{map}
\title{Velicer MAP Test}
\usage{
map(
  data,
  vars,
  type = "pearson",
  eigen = TRUE,
  avgCorr = FALSE,
  screePlot = FALSE,
  MapCurvePlot = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{type}{.}

\item{eigen}{.}

\item{avgCorr}{.}

\item{screePlot}{.}

\item{MapCurvePlot}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$eigen} \tab \tab \tab \tab \tab a table \cr
  \code{results$avgCorr} \tab \tab \tab \tab \tab a table \cr
  \code{results$screePlot} \tab \tab \tab \tab \tab an image \cr
  \code{results$MapCurvePlot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$eigen$asDF}

\code{as.data.frame(results$eigen)}
}
\description{
Velicer MAP Test
}
