% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcomp.h.R
\name{netcomp}
\alias{netcomp}
\title{Group Network Comparison}
\usage{
netcomp(
  data,
  vars,
  group,
  run = FALSE,
  nPerm = 200,
  testStructure = FALSE,
  testGlobalStrength = FALSE,
  testEdges = FALSE,
  showPlots = TRUE,
  layout = "spring",
  sameLayout = TRUE,
  minimum = 0
)
}
\arguments{
\item{data}{.}

\item{vars}{.}

\item{group}{.}

\item{run}{.}

\item{nPerm}{.}

\item{testStructure}{.}

\item{testGlobalStrength}{.}

\item{testEdges}{.}

\item{showPlots}{.}

\item{layout}{.}

\item{sameLayout}{.}

\item{minimum}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$groupPlots$plotGroup1} \tab \tab \tab \tab \tab an image \cr
  \code{results$groupPlots$plotGroup2} \tab \tab \tab \tab \tab an image \cr
  \code{results$summary} \tab \tab \tab \tab \tab a table \cr
  \code{results$tests} \tab \tab \tab \tab \tab a table \cr
  \code{results$edges} \tab \tab \tab \tab \tab a table \cr
  \code{results$note} \tab \tab \tab \tab \tab a html \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$summary$asDF}

\code{as.data.frame(results$summary)}
}
\description{
Group Network Comparison
}
