% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.h.R
\name{network}
\alias{network}
\title{Directed Graph Centrality}
\usage{
network(
  data,
  labels,
  vars,
  cen = TRUE,
  plot = FALSE,
  absWeight = "TRUE",
  threshold = 0
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{labels}{.}

\item{vars}{.}

\item{cen}{.}

\item{plot}{.}

\item{absWeight}{.}

\item{threshold}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$filterInfoTable} \tab \tab \tab \tab \tab a html \cr
  \code{results$cen} \tab \tab \tab \tab \tab a table \cr
  \code{results$filterInfoPlot} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$cen$asDF}

\code{as.data.frame(results$cen)}
}
\description{
Directed Graph Centrality
}
