% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.h.R
\name{poly}
\alias{poly}
\title{Polychoric Correlation}
\usage{
poly(
  data,
  vars,
  polychoric = TRUE,
  scale = "raw",
  plot = FALSE,
  plot2 = FALSE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{polychoric}{.}

\item{scale}{.}

\item{plot}{.}

\item{plot2}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$matrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$matrix$asDF}

\code{as.data.frame(results$matrix)}
}
\description{
Polychoric Correlation
}
