% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.h.R
\name{rank}
\alias{rank}
\title{Ordinal Correlation}
\usage{
rank(
  data,
  vars,
  type = "spearman",
  scale = "raw",
  cor = TRUE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{type}{.}

\item{scale}{.}

\item{cor}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$matrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$matrix$asDF}

\code{as.data.frame(results$matrix)}
}
\description{
Ordinal Correlation
}
