% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsurvey.h.R
\name{ahpsurvey}
\alias{ahpsurvey}
\title{AHP for survey data}
\usage{
ahpsurvey(
  data,
  vars,
  atts = "cult, fam, house, jobs, trans",
  itemmat = TRUE,
  weights = FALSE,
  cir = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{atts}{.}

\item{itemmat}{.}

\item{weights}{.}

\item{cir}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$itemmat} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$itemmat$asDF}

\code{as.data.frame(results$itemmat)}
}
\description{
AHP for survey data
}
