% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsurvey.h.R
\name{ahpsurvey}
\alias{ahpsurvey}
\title{AHP for survey data}
\usage{
ahpsurvey(data, vars, atts = "cult, fam, house, jobs, trans", plot1 = TRUE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{atts}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
AHP for survey data
}
