% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.h.R
\name{concordance}
\alias{concordance}
\title{Concordance Correlation}
\usage{
concordance(
  data,
  dep,
  covs,
  cc = TRUE,
  ccp = FALSE,
  bap = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{covs}{.}

\item{cc}{.}

\item{ccp}{.}

\item{bap}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$table} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$table$asDF}

\code{as.data.frame(results$table)}
}
\description{
Concordance Correlation
}
