% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.h.R
\name{cor}
\alias{cor}
\title{Correlation structure}
\usage{
cor(
  data,
  vars,
  type = "pearson",
  method = "none",
  k = 0,
  size = 3,
  plot = TRUE,
  horiz = FALSE,
  plot1 = FALSE,
  poly = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  horiz1 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{type}{.}

\item{method}{.}

\item{k}{.}

\item{size}{.}

\item{plot}{.}

\item{horiz}{.}

\item{plot1}{.}

\item{poly}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{horiz1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Correlation structure
}
