% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.h.R
\name{cor}
\alias{cor}
\title{Correlation Structure}
\usage{
cor(
  data,
  vars,
  type = "pearson",
  mat = TRUE,
  missing = "pairwise.complete.obs",
  order = "hclust",
  method = "complete",
  k = 2,
  size = 3,
  color = "blue",
  plot = FALSE,
  method1 = "circle",
  type1 = "full",
  width = 500,
  height = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{type}{.}

\item{mat}{.}

\item{missing}{.}

\item{order}{.}

\item{method}{.}

\item{k}{.}

\item{size}{.}

\item{color}{.}

\item{plot}{.}

\item{method1}{.}

\item{type1}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$matrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$matrix$asDF}

\code{as.data.frame(results$matrix)}
}
\description{
Correlation Structure
}
