% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.h.R
\name{kappa}
\alias{kappa}
\title{Fleiss Kappa and Krippendorff alpha}
\usage{
kappa(
  data,
  vars,
  fk = TRUE,
  ek = TRUE,
  cw = FALSE,
  krip = FALSE,
  method = "nominal"
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{fk}{.}

\item{ek}{.}

\item{cw}{.}

\item{krip}{.}

\item{method}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fk} \tab \tab \tab \tab \tab a table \cr
  \code{results$ek} \tab \tab \tab \tab \tab a table \cr
  \code{results$cw} \tab \tab \tab \tab \tab a table \cr
  \code{results$krip} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fk$asDF}

\code{as.data.frame(results$fk)}
}
\description{
Fleiss Kappa and Krippendorff alpha
}
