% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.h.R
\name{multilevel}
\alias{multilevel}
\title{Multilevel Correlation}
\usage{
multilevel(data, vars, facs, icc = TRUE, multi = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{facs}{.}

\item{icc}{.}

\item{multi}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$icc} \tab \tab \tab \tab \tab a table \cr
  \code{results$multi} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$icc$asDF}

\code{as.data.frame(results$icc)}
}
\description{
Multilevel Correlation
}
