% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.h.R
\name{partial}
\alias{partial}
\title{Partial Correlation}
\usage{
partial(
  data,
  vars,
  ctrlvars,
  sidSig = "twotailed",
  scale = "raw",
  shwSig = FALSE,
  flgSig = FALSE,
  plot = TRUE,
  model = "glasso",
  layout = "spring",
  shape = "circle",
  plot2 = FALSE,
  ebic = FALSE,
  pm = FALSE,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{ctrlvars}{.}

\item{sidSig}{.}

\item{scale}{.}

\item{shwSig}{.}

\item{flgSig}{.}

\item{plot}{.}

\item{model}{.}

\item{layout}{.}

\item{shape}{.}

\item{plot2}{.}

\item{ebic}{.}

\item{pm}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$matrix} \tab \tab \tab \tab \tab correlation matrix table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$matrix$asDF}

\code{as.data.frame(results$matrix)}
}
\description{
Partial Correlation
}
