% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater.h.R
\name{rater}
\alias{rater}
\title{Rater Reliability}
\usage{
rater(
  data,
  vars,
  model = "oneway",
  type = "agreement",
  unit = "single",
  pair = FALSE,
  interrater = FALSE,
  icc = FALSE,
  bicc = FALSE,
  ic = FALSE,
  ftest = FALSE,
  fk = FALSE,
  cw = FALSE,
  ek = FALSE,
  krip = FALSE,
  method = "nominal",
  t = "col",
  pa = FALSE,
  boot = 1000,
  bt = FALSE,
  boot1 = 1000,
  kend = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{model}{.}

\item{type}{.}

\item{unit}{.}

\item{pair}{.}

\item{interrater}{.}

\item{icc}{.}

\item{bicc}{.}

\item{ic}{.}

\item{ftest}{.}

\item{fk}{.}

\item{cw}{.}

\item{ek}{.}

\item{krip}{.}

\item{method}{.}

\item{t}{.}

\item{pa}{.}

\item{boot}{.}

\item{bt}{.}

\item{boot1}{.}

\item{kend}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$matrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$interrater} \tab \tab \tab \tab \tab a table \cr
  \code{results$fk} \tab \tab \tab \tab \tab a table \cr
  \code{results$bt} \tab \tab \tab \tab \tab a table \cr
  \code{results$ek} \tab \tab \tab \tab \tab a table \cr
  \code{results$cw} \tab \tab \tab \tab \tab a table \cr
  \code{results$pa} \tab \tab \tab \tab \tab a table \cr
  \code{results$icc} \tab \tab \tab \tab \tab a table \cr
  \code{results$bicc} \tab \tab \tab \tab \tab a table \cr
  \code{results$kend} \tab \tab \tab \tab \tab a table \cr
  \code{results$ic} \tab \tab \tab \tab \tab a table \cr
  \code{results$ftest} \tab \tab \tab \tab \tab a table \cr
  \code{results$krip} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$matrix$asDF}

\code{as.data.frame(results$matrix)}
}
\description{
Rater Reliability
}
