% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater.h.R
\name{rater}
\alias{rater}
\title{Rater Reliability}
\usage{
rater(data, vars, interrater = TRUE, icc = TRUE, bicc = FALSE, ggm = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{interrater}{.}

\item{icc}{.}

\item{bicc}{.}

\item{ggm}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$interrater} \tab \tab \tab \tab \tab a table \cr
  \code{results$icc} \tab \tab \tab \tab \tab a table \cr
  \code{results$bicc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$interrater$asDF}

\code{as.data.frame(results$interrater)}
}
\description{
Rater Reliability
}
