% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rater.h.R
\name{rater}
\alias{rater}
\title{Rater Reliability}
\usage{
rater(
  data,
  vars,
  model = "oneway",
  type = "agreement",
  unit = "single",
  interrater = TRUE,
  icc = FALSE,
  bicc = FALSE,
  ic = TRUE,
  ftest = FALSE,
  ggm = FALSE,
  par = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{model}{.}

\item{type}{.}

\item{unit}{.}

\item{interrater}{.}

\item{icc}{.}

\item{bicc}{.}

\item{ic}{.}

\item{ftest}{.}

\item{ggm}{.}

\item{par}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$interrater} \tab \tab \tab \tab \tab a table \cr
  \code{results$icc} \tab \tab \tab \tab \tab a table \cr
  \code{results$bicc} \tab \tab \tab \tab \tab a table \cr
  \code{results$ic} \tab \tab \tab \tab \tab a table \cr
  \code{results$ftest} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$interrater$asDF}

\code{as.data.frame(results$interrater)}
}
\description{
Rater Reliability
}
