% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmc.h.R
\name{rmc}
\alias{rmc}
\title{Repeated Measures Correlation}
\usage{
rmc(data, id, dep, covs, cc = TRUE, plot = FALSE, width = 500, height = 500)
}
\arguments{
\item{data}{.}

\item{id}{.}

\item{dep}{.}

\item{covs}{.}

\item{cc}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$cc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$cc$asDF}

\code{as.data.frame(results$cc)}
}
\description{
Repeated Measures Correlation
}
