% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmc.h.R
\name{rmc}
\alias{rmc}
\title{Repeated & Cross Correlation}
\usage{
rmc(
  data,
  id,
  dep,
  covs,
  rc = FALSE,
  plot = FALSE,
  width = 500,
  height = 500,
  cc = FALSE,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500
)
}
\arguments{
\item{data}{.}

\item{id}{.}

\item{dep}{.}

\item{covs}{.}

\item{rc}{.}

\item{plot}{.}

\item{width}{.}

\item{height}{.}

\item{cc}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$rc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$cc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rc$asDF}

\code{as.data.frame(results$rc)}
}
\description{
Repeated & Cross Correlation
}
