% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmc.h.R
\name{rmc}
\alias{rmc}
\title{Repeated & Cross Correlation}
\usage{
rmc(data, id, dep, covs, rc = FALSE, plot = FALSE, cc = FALSE, plot1 = FALSE)
}
\arguments{
\item{data}{.}

\item{id}{.}

\item{dep}{.}

\item{covs}{.}

\item{rc}{.}

\item{plot}{.}

\item{cc}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$rc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$cc} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rc$asDF}

\code{as.data.frame(results$rc)}
}
\description{
Repeated & Cross Correlation
}
